Remove trace of interrupt. Add timeoit while waiting for ECHO pin. Use copy_from_user() rather than direct access to user-space pointer. Add code to auto-create vfs node in /dev.

+++ driver-recipe/hcsr04.h
@@ -20,6 +20,6 @@
 */
 struct hcsr04_data 
 {
-    unsigned long long t_stamp;
-    int t_high;
+    unsigned long t_stamp;
+    unsigned int t_high;
 };
